Skip incorrect migrations in new strangelog-projects
**Steps to reproduce:**
1. Setup a new strangelog project (this means to just install strangelog
   as a dependency since ther eis no "init"-step)
2. Add an entry (e.g. via `strangelog add`)
3. Run `strangelog migrate`

**Expected:**
See "Migrated from [LATEST] to [LATEST]" (since this is a new project and
there should be nothing to migrate)

**Actual:**
- see "Migrated from -1 to [LATEST]"
- have broken version directory names (`1.0.0` becomes `1.0.0.0`)

**Technical Reason:**
The absence of the `info.yml` is interpreted as a project being on the
migration version `-1` of strangelog (the version before `info.yml` and thus
migrations were introduced). In new projects however, an `info.yml` also does
not exists due to the (intentional) lack of an "init"-command/step.

**Fix:**
Differentiate between
- a project not having an `info.yml` and no entry files yet (assumes a new,
  already "migrated" project)
- a project not having an `info.yml` but some entry files already (assumes
  an old, not yet migrated project)